# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Sistem Manajemen Parkir.",
        "title": "Parking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
      "/api/v1/admin/reports/daily-settlements": {
        "get": {
          "description": "Retrieve all daily settlements with pagination support",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Reporting"
          ],
          "summary": "Get all daily settlements",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "description": "Search term",
              "name": "search",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "description": "Create a new daily settlement record",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Reporting"
          ],
          "summary": "Create a new daily settlement",
          "parameters": [
            {
              "description": "Daily settlement data",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateDailySettlementRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/api/v1/admin/reports/daily-settlements/{id}": {
        "get": {
          "description": "Retrieve a specific daily settlement by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Reporting"
          ],
          "summary": "Get daily settlement by ID",
          "parameters": [
            {
              "type": "string",
              "description": "Daily settlement ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "description": "Delete a specific daily settlement by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Reporting"
          ],
          "summary": "Delete a daily settlement",
          "parameters": [
            {
              "type": "string",
              "description": "Daily settlement ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/api/v1/admin/reports/daily-settlements/{id}/reconcile": {
        "post": {
          "description": "Reconcile a specific daily settlement by its ID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Reporting"
          ],
          "summary": "Reconcile a daily settlement",
          "parameters": [
            {
              "type": "string",
              "description": "Daily settlement ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Reconciliation data",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.ReconcileDailySettlementRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
...
      "/customers": {
        "get": {
          "description": "Get list of customers with pagination",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Customers"
          ],
          "summary": "Get All Customers",
          "parameters": [
            {
              "type": "integer",
              "default": 1,
              "description": "Page number",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "default": 10,
              "description": "Items per page",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "description": "Search by name, phone, or rfid_uid",
              "name": "search",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.PageResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "post": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Customers"
          ],
          "summary": "Create New Customer",
          "parameters": [
            {
              "description": "Customer Request",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.CreateCustomerRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
      "/customers/{id}": {
        "get": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Customers"
          ],
          "summary": "Get Customer By ID",
          "parameters": [
            {
              "type": "string",
              "description": "Customer Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "404": {
              "description": "Not Found",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "put": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Customers"
          ],
          "summary": "Update Customer",
          "parameters": [
            {
              "type": "string",
              "description": "Customer Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "description": "Update Request Body",
              "name": "request",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/requests.UpdateCustomerRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        },
        "delete": {
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Customers"
          ],
          "summary": "Delete Customer",
          "parameters": [
            {
              "type": "string",
              "description": "Customer Hash ID",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/responses.BaseResponse"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "requests.CreateCustomerRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "registration_source_id": {
            "type": "string"
          },
          "rfid_uid": {
            "type": "string",
            "maxLength": 50
          }
        }
      },
      "requests.CreateDailySettlementRequest": {
        "type": "object",
        "required": [
          "date",
          "total_cash",
          "total_lost_tickets",
          "total_qris",
          "total_refunds",
          "total_revenue",
          "total_transactions"
        ],
        "properties": {
          "date": {
            "type": "string"
          },
          "total_cash": {
            "type": "number",
            "minimum": 0
          },
          "total_lost_tickets": {
            "type": "integer",
            "minimum": 0
          },
          "total_qris": {
            "type": "number",
            "minimum": 0
          },
          "total_refunds": {
            "type": "number",
            "minimum": 0
          },
          "total_revenue": {
            "type": "number",
            "minimum": 0
          },
          "total_transactions": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "requests.ReconcileDailySettlementRequest": {
        "type": "object",
        "required": [
          "reconciled_by"
        ],
        "properties": {
          "notes": {
            "type": "string",
            "maxLength": 500
          },
          "reconciled_by": {
            "type": "string"
          }
        }
      },
...
      "requests.UpdateCustomerRequest": {
        "type": "object",
        "properties": {
          "is_registered": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "registration_source_id": {
            "type": "string"
          },
          "rfid_uid": {
            "type": "string",
            "maxLength": 50
                }
            }
        },
...
        "responses.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "current_url": {
                    "description": "Navigasi URL",
                    "type": "string"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "first_page": {
                    "description": "Navigasi Angka",
                    "type": "integer"
                },
                "first_url": {
                    "type": "string"
                },
                "last_page": {
                    "type": "integer"
                },
                "last_url": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "next_url": {
                    "type": "string"
                },
                "prev_page": {
                    "description": "Pakai pointer biar bisa null kalau gak ada",
                    "type": "integer"
                },
                "prev_url": {
                    "type": "string"
                },
                "search": {
                    "type": "string"
                },
                "sort": {
                    "description": "Search \u0026 Filter",
                    "type": "string"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "responses.PageResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/responses.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}